+++ tests/src/Interop/PInvoke/SizeParamIndex/ReversePInvoke/PassingByRef/ReversePInvokePassingByRefNative.cpp
@@ -4,7 +4,7 @@
 
 // ReversePInvokePassingByRefNative.cpp : Defines the entry point for the DLL application.
 #include <xplatform.h>
-#include "../helper.h"
+#include "helper.h"
 
 //Func Pointer
 typedef BOOL (__cdecl *DelByteArrByRefAsCdeclCaller)(BYTE** arrByte, BYTE* arraySize);
@@ -41,7 +41,7 @@ extern "C" DLL_EXPORT BOOL __cdecl DoCallBack_MarshalByteArray_AsParam_AsByRef(D
 extern "C" DLL_EXPORT BOOL __cdecl DoCallBack_MarshalSbyteArray_AsParam_AsByRef(DelSbyteArrByRefAsCdeclCaller caller)
 {
     CHAR arrSize = 1;
-    CHAR* arrSbyte = InitArray<CHAR>(arrSize);
+    CHAR* arrSbyte = InitArray<CHAR>((size_t)arrSize);
 
     if(!caller(&arrSize, &arrSbyte))
     {
@@ -50,7 +50,7 @@ extern "C" DLL_EXPORT BOOL __cdecl DoCallBack_MarshalSbyteArray_AsParam_AsByRef(
         return FALSE;
     }
 
-    return CheckArray(arrSbyte, arrSize, (CHAR)127);
+    return CheckArray(arrSbyte, (size_t)arrSize, (CHAR)127);
 }
 
 //SHORT -1 ==> 20 size Array(Actual: 10 ==> 20)
@@ -66,14 +66,14 @@ extern "C" DLL_EXPORT BOOL __cdecl DoCallBack_MarshalShortArray_AsParam_AsByRef(
         return FALSE;
     }
 
-    return CheckArray(arrShort, arrSize, (SHORT)20);
+    return CheckArray(arrShort, (size_t)arrSize, (SHORT)20);
 }
 
 //SHORT 10 ==> -1 size Array(Actual: 10 ==> 20)
 extern "C" DLL_EXPORT BOOL __cdecl DoCallBack_MarshalShortArrayReturnNegativeSize_AsParam_AsByRef(DelShortArrByRefAsCdeclCaller caller)
 {
     SHORT arrSize = 10;
-    SHORT* arrShort = InitArray<SHORT>(arrSize);
+    SHORT* arrShort = InitArray<SHORT>((size_t)arrSize);
 
     if(!caller(&arrShort, &arrSize))
     {
@@ -108,7 +108,7 @@ extern "C" DLL_EXPORT BOOL __cdecl DoCallBack_MarshalUshortArray_AsParam_AsByRef
 extern "C" DLL_EXPORT BOOL __cdecl DoCallBack_MarshalInt32Array_AsParam_AsByRef(DelInt32ArrByRefAsCdeclCaller caller)
 {
     LONG arrSize = 10;
-    LONG* arrInt32 = InitArray<LONG>(arrSize);
+    LONG* arrInt32 = InitArray<LONG>((size_t)arrSize);
 
     if(!caller(&arrInt32, &arrSize))
     {
@@ -117,7 +117,7 @@ extern "C" DLL_EXPORT BOOL __cdecl DoCallBack_MarshalInt32Array_AsParam_AsByRef(
         return FALSE;
     }
 
-    return CheckArray(arrInt32, arrSize, (LONG)20);
+    return CheckArray(arrInt32, (size_t)arrSize, (LONG)20);
 }
 
 //UInt32 10 ==> 20 size Array
